requirements.txt
Django==4.2
django-cors-headers==4.5.0
django-countries==7.6.1
django-filter==23.5
djangorestframework_gis==1.1
djangorestframework==3.14
psycopg2==2.9.10
PyJWT==1.7.1
python_magic==0.4.15
pyyaml==5.1
